implement validation

+++ prescription-generator/src/main/java/com/example/prescription_generator/controller/PrescriptionController.java
@@ -4,6 +4,8 @@ import com.example.prescription_generator.model.dto.PrescriptionDTO;
 import com.example.prescription_generator.model.entity.Prescription;
 import com.example.prescription_generator.model.mapper.PrescriptionMapper;
 import com.example.prescription_generator.service.PrescriptionService;
+import com.example.prescription_generator.service.PrescriptionValidationService;
+import com.example.prescription_generator.service.ValidationService;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -20,21 +22,31 @@ import java.util.Map;
 public class PrescriptionController {
     private final PrescriptionService prescriptionService;
     private final PrescriptionMapper prescriptionMapper;
+    private final ValidationService validationService;
+    private final PrescriptionValidationService prescriptionValidationService;
 
     @PostMapping("/register")
     public ResponseEntity<?> register(@RequestBody PrescriptionDTO prescriptionDTO){
-
+        if(!validationService.validatePrescriptionDetails(prescriptionDTO).isEmpty()){
+            return ResponseEntity.badRequest().body(validationService.validatePrescriptionDetails(prescriptionDTO));
+        }
         return ResponseEntity.ok(prescriptionMapper.toPrescriptionDTO(
                 prescriptionService.savePrescription(prescriptionMapper.toPrescription(prescriptionDTO))));
     }
     @PutMapping("/update/{id}")
     public ResponseEntity<?> update(@RequestBody PrescriptionDTO prescriptionDTO, @PathVariable("id") Long id){
+        if(!validationService.validatePrescriptionDetails(prescriptionDTO).isEmpty()){
+            return ResponseEntity.badRequest().body(validationService.validatePrescriptionDetails(prescriptionDTO));
+        }
         return ResponseEntity.ok(prescriptionMapper.toPrescriptionDTO(
                 prescriptionService.updatePrescription(id,prescriptionDTO)
         ));
     }
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> delete(@PathVariable("id") Long id){
+        if (!prescriptionValidationService.isExitPrescriptionById(id)){
+            return ResponseEntity.badRequest().body("Prescription doesn't exit!");
+        }
         prescriptionService.deletePrescription(id);
         return ResponseEntity.ok("Prescription deleted successfully!");
     }
